requirements.txt
# Core dependencies
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
requests>=2.31.0
PyYAML>=6.0
